file system initial design
