+++ pyproject.toml
@@ -7,7 +7,7 @@ dependencies = [
 description = "Triple-Axis Data Visualization Toolkit"
 dynamic = ["version"]
 license = "LGPL-3.0-or-later"
-name = "neutron-tavi"
+name = "tavi"
 requires-python = ">=3.10"
 
 [project.urls]
@@ -47,8 +47,8 @@ where = ["src"]
 # [project.scripts]
 # packagename-cli = "packagenamepy.packagename:main"
 
-# [project.gui-scripts]
-# packagenamepy = "packagenamepy.packagename:gui"
+[project.gui-scripts]
+tavi = "tavi.tavimain:gui"
 
 [tool.pytest.ini_options]
 markers = [
