+++ src/tavi/_version.py
@@ -1,0 +1,1 @@
+__version__ = "0.3.0.dev25"
