+++ src/tavi/tavi_model/FileSystem/tavi_data_schema.json
@@ -1,0 +1,93 @@
+{
+  "RawMetaData": {
+        "scan": "str",
+        "date": "str",
+        "time": "str",
+        "proposal": "str",
+        "experiment": "str",
+        "experiment_number": "str",
+        "command": "str",
+        "builtin_command": "str",
+        "users": "str",
+        "local_contact": "str",
+        "scan_title": "str",
+        "monochromator": "str",
+        "analyzer": "str",
+        "sense": "str",
+        "collimation": "str",
+        "samplename": "str",
+        "sampletype": "str",
+        "samplemosaic": "str",
+        "latticeconstants": "str",
+        "ubmatrix": "str",
+        "mode": "str",
+        "plane_normal": "str",
+        "ubconf": "str",
+        "preset_type": "str",
+        "preset_channel": "str",
+        "preset_value": "str",
+        "def_x": "str",
+        "def_y": "str",
+        "sum_of_counts": "str",
+        "center_of_mass": "str",
+        "full_width_half_maximum": "str",
+        "end_time": "str"
+    },
+  "RawData": {
+    "Pt": "ndarray",
+    "s1": "ndarray",
+    "time": "ndarray",
+    "detector": "ndarray",
+    "monitor": "ndarray",
+    "mcu": "ndarray",
+    "focal_length": "ndarray",
+    "m1": "ndarray",
+    "m2": "ndarray",
+    "marc": "ndarray",
+    "mtrans": "ndarray",
+    "mfocus": "ndarray",
+    "s2": "ndarray",
+    "sgl": "ndarray",
+    "sgu": "ndarray",
+    "bbb": "ndarray",
+    "bbl": "ndarray",
+    "bbr": "ndarray",
+    "bbt": "ndarray",
+    "bab": "ndarray",
+    "bal": "ndarray",
+    "bar": "ndarray",
+    "bat": "ndarray",
+    "stl": "ndarray",
+    "stu": "ndarray",
+    "a1": "ndarray",
+    "a2": "ndarray",
+    "afocus": "ndarray",
+    "qm1": "ndarray",
+    "qm2": "ndarray",
+    "qm3": "ndarray",
+    "qm4": "ndarray",
+    "qm5": "ndarray",
+    "qm6": "ndarray",
+    "qm7": "ndarray",
+    "qm8": "ndarray",
+    "xm1": "ndarray",
+    "xm2": "ndarray",
+    "xm3": "ndarray",
+    "xm4": "ndarray",
+    "xm5": "ndarray",
+    "xm6": "ndarray",
+    "xm7": "ndarray",
+    "xm8": "ndarray",
+    "q": "ndarray",
+    "h": "ndarray",
+    "k": "ndarray",
+    "l": "ndarray",
+    "ei": "ndarray",
+    "ef": "ndarray",
+    "e": "ndarray",
+    "vti": "ndarray",
+    "sample": "ndarray",
+    "temp": "ndarray",
+    "temp_2": "ndarray"
+  }
+}
